% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Example title is an incomplete sentence --no period}
\usage{
dummy(x, y = 1L)
}
\arguments{
\item{x}{Integer to increment.  Required.}

\item{y}{Amount to increment.  Defaults to 1. Required integer.}
}
\value{
A \link[base:integer]{base::integer} that is incremented by \code{y}.
}
\description{
"Description" is a series of complete sentences.
See https://r-pkgs.org/man.html.
Keep these documentation lines under 80 or 100 characters
(I forgot which )
}
\details{
Adds one.
}
\note{
Use "note" for something more obscure than a "description"?
}
\examples{
dummy(0L)
dummy(1L)
dummy(0L, 2L)

}
\author{
Yutian Thompson
}
