% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Example title is an incomplete sentence --no period}
\usage{
dummy(x, y = 1L)
}
\arguments{
\item{x}{Integer to increment Required.}

\item{y}{Amount to increment.  Defaults to 1. Required.}
}
\value{
An \link[base:integer]{base::integer}
}
\description{
"Description" is a series of complete sentence
}
\details{
Adds one.
}
\note{
Use "note" for something more obscure than a "description"?
}
\examples{
dummy(0L)
dummy(1L)
dummy(0L, 2L)
}
\author{
Yutian Thompson
}
